+++ docs/release-notes.md
@@ -1,6 +1,7 @@
 # New features and Changes
 
 - Add multipart file download
+- Add custom anime specific folders
 - Improve anime folder detection
 - Add close minimizes to tray setting